Remind again if the notification is reposted even if ignored before
This helps to handle cases where messaging apps such as Hangouts bundle
multiple messages into the same notification. Although user has ignored the
notification, they may still want to be notified on further messages.
Unfortunately this is only possible on v18, because on v14 the
AccessibilityEvent.TYPE_NOTIFICATION_STATE_CHANGED event is not sent when
notification changes, therefore we have not way of detecting that.